+++ xla/stream_executor/cuda/BUILD
@@ -16,10 +16,6 @@ load(
     "if_cuda_is_configured",
 )
 load(
-    "@local_config_rocm//rocm:build_defs.bzl",
-    "if_rocm_is_configured",
-)
-load(
     "@tsl//tsl/platform:build_config_root.bzl",
     "if_static",
     "tf_cuda_tests_tags",
@@ -251,16 +247,7 @@ cc_library(
         "//xla/stream_executor:host_or_device_scalar",
         "//xla/stream_executor:stream_executor_headers",
         "//xla/stream_executor/platform",
-    ]) + if_rocm_is_configured([
-        "@local_config_rocm//rocm:rocm_headers",
     ]) + ["@tsl//tsl/platform:errors"],
-)
-
-cc_library(
-    name = "cublas_lt_header_for_rocm",
-    hdrs = ["cuda_blas_lt_impl.h",
-            "cuda_blas_lt.h"],
-    visibility = ["//visibility:public"],
 )
 
 alias(
@@ -284,7 +271,6 @@ cc_library(
     hdrs = if_cuda_is_configured([
         "cuda_blas.h",
         "cuda_blas_lt.h",
-        "cuda_blas_lt_impl.h",
     ]),
     visibility = ["//visibility:public"],
     deps = if_cuda_is_configured([
